Revert "Merge remote-tracking branch 'refs/remotes/origin/development' into to_master_wotg_development_merge"
This reverts commit 28b6ce31f728a9f20eca1255473ed3ccd6c4c92e [formerly 3379359a5f00168dadb87cb0ebda436756daddcd], reversing
changes made to a420122769739b02e8731dc9ee7e517d753a8461 [formerly 57ed3298ba71bba7393c4e408966a9467305aa02].


Former-commit-id: ce7404d0933b64de196fa175f96173881371ab90

+++ Projects/UnitTests/Classes/UIListTest.cpp
@@ -76,14 +76,14 @@ UIListCell *UIListTestDelegate::CellAtIndex(UIList *list, int32 index)
     return cell;
 }
 
-float32 UIListTestDelegate::CellWidth(UIList*, int32)
+int32 UIListTestDelegate::CellWidth(UIList*, int32)
 {
-   	return cellSize.x;
+   	return (int32)cellSize.x;
 }
 
-float32 UIListTestDelegate::CellHeight(UIList*, int32)
+int32 UIListTestDelegate::CellHeight(UIList*, int32)
 {
-   	return cellSize.y;
+   	return (int32)cellSize.y;
 }
 
 UIListTest::UIListTest() :
